ref(console-login): Extract the encryption section into a separate utility class
